including some figures

+++ simulations/vary_all/norm/R-ind/norm_pairwise_N01.R
@@ -14,7 +14,8 @@ set.seed(9999)
 nsim = 200
 # change path while running it on HPC
 # simtable<-read_csv(here::here('simulations/null/sim_table.csv'))
-simtable<-read_csv('../../../sim_table/sim_table.csv')
+simtable<-read_csv('../sim_table.csv')
+
 
 ### Extract flags from simulation scenario
 
@@ -24,20 +25,28 @@ scen<-as.numeric(commandArgs()[[6]]) # If running batch job uncomment this
 simj<-simtable[scen,] #Extract row of table
 nfacetj<-simj$nfacet # Which nfacet level
 nxj<-simj$nx #Which nx level
+wj <- simj$w
 
 #create data for each row for null normal
 
-sim_null_normal = function(nxj, nfacetj){
-  rep(distributional::dist_normal(0, 1), 
-      times = nxj*nfacetj)
+sim_varall = function(nx,
+                      nfacet,
+                      mean = 0,
+                      sd = 1,
+                      w)
+{
+  dist_normal((mean + seq(0, (nx*nfacet - 1), by  = 1)*w), sd)
 }
+
 
 sim_panel_data = 
   hakear::sim_panel(nx = nxj,
                     nfacet = nfacetj, 
                     ntimes = 500, 
-                    sim_dist = sim_null_normal) %>% 
-  unnest(c(data)) %>% ungroup()
+                    sim_dist = sim_varall(nx = nxj, nfacet = nfacetj, w = wj)) %>% 
+  unnest(c(data)) %>%
+  ungroup()%>% 
+  bind_cols(w = wj)
 
 set.seed(1111)
 
@@ -58,7 +67,13 @@ set.seed(1111)
   
 }) %>% bind_rows()
 
-saveRDS(raw_dist,
+if(!dir.exists("simulations/vary_all/norm/data-ind/wpd_N01"))
+  {
+    dir.create("simulations/vary_all/norm/data-ind/wpd_N01")
+  }  
+  
+  saveRDS(raw_dist,
           paste0('../data-ind/wpd_N01/',
                  nxj,'_',
-                 nfacetj,'_wpd.rds'))
+                 nfacetj,"_",
+                 wj, '_wpd.rds'))
